Added a rank system that is stored in the database
Each rank can be defined in the configuration file along with it's required XP. The player rank and currentXP is store in the database.
Eventually when a player can gain XP it'll add to the database currentXP value and then calculate if they deserve a rankup, therefore allowing them access to more kits.