Reworked teams and grouping players
Teams are now initialised upon the server starting, allowing the ability to store players in the team using their SteamID. This allows me to get a team player count easily.

Created a AssignPlayerToGroup method to reduce the amount of code duplication and makes it more modular.

Still need to work on a balancing system, and add the ability for players to change teams.