Added randomised outfits
Created a clear clothing method based on the clear inventory method that resets all the player's clothing to nothing.
Created a Outfit model that allows the creation of outfits based on an array of clothing IDs.
Created methods for when the player spawns and returns to lobby to give them the appropriate clothing. Upon spawning they get a random outfit from the list of outfits, but when they return to the lobby they get given a generic lobby outfit.